Fix compiler warnings: CS0414, CS0618, CS0108, CS0219, CS0168, CS0169, CS0675

+++ ManicDiggerLib/Server/Server.cs
@@ -4096,7 +4096,7 @@ public static class MapUtil
         ulong v = 0;
         v = (ulong)x << 40;
         v |= (ulong)y << 20;
-        v |= (ulong)z;
+        v |= (ulong)(uint)z;
         return v;
     }
